requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Image Processing
Pillow==10.1.0
numpy==1.24.3

# Medical AI Models (Optional - for real AI features)
# Uncomment these if you want real AI capabilities:
torch==2.1.0
torchvision==0.16.0
torchxrayvision==1.2.0
easyocr==1.7.0
scikit-image==0.22.0  # Required for image preprocessing

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Environment Management
python-dotenv==1.0.0

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools (Optional)
black==23.11.0
flake8==6.1.0
